+++ src/Meta.js
@@ -1,3 +1,0 @@
-module.exports = {
-	version: '1.3.19'
-};